+++ src/components/modal/modal.e2e.ts
@@ -23,7 +23,9 @@ describe("calcite-modal properties", () => {
 
   it("sets custom width correctly", async () => {
     const page = await newE2EPage();
-    await page.setContent(`<calcite-modal width="400"></calcite-modal>`);
+    // set large page to ensure test modal isn't becoming fullscreen
+    page.setViewport({ width: 1440, height: 1440 });
+    await page.setContent(`<calcite-modal style="--calcite-modal-width:600px;"></calcite-modal>`);
     const modal = await page.find("calcite-modal");
     await modal.setProperty("open", true);
     await page.waitForChanges();
@@ -31,7 +33,74 @@ describe("calcite-modal properties", () => {
       const m = elm.shadowRoot.querySelector(".modal");
       return window.getComputedStyle(m).getPropertyValue("width");
     });
-    expect(style).toEqual("400px");
+    expect(style).toEqual("600px");
+  });
+
+  it("sets custom height correctly", async () => {
+    const page = await newE2EPage();
+    // set large page to ensure test modal isn't becoming fullscreen
+    page.setViewport({ width: 1440, height: 1440 });
+    await page.setContent(`<calcite-modal style="--calcite-modal-height:600px;" open></calcite-modal>`);
+    const modal = await page.find("calcite-modal");
+    await modal.setProperty("open", true);
+    await page.waitForChanges();
+    const style = await page.$eval("calcite-modal", (elm) => {
+      const m = elm.shadowRoot.querySelector(".modal");
+      return window.getComputedStyle(m).getPropertyValue("height");
+    });
+    expect(style).toEqual("600px");
+  });
+
+  it("expectedly does not set custom width when `fullscreen` is true", async () => {
+    const page = await newE2EPage();
+    // set large page to ensure test modal isn't becoming fullscreen
+    page.setViewport({ width: 1440, height: 1440 });
+    await page.setContent(`<calcite-modal style="--calcite-modal-width:600px;" fullscreen open></calcite-modal>`);
+    const modal = await page.find("calcite-modal");
+    await modal.setProperty("open", true);
+    await page.waitForChanges();
+    const style = await page.$eval("calcite-modal", (elm) => {
+      const m = elm.shadowRoot.querySelector(".modal");
+      return window.getComputedStyle(m).getPropertyValue("width");
+    });
+    expect(style).not.toEqual("600px");
+  });
+
+  it("expectedly does not set custom height when `fullscreen` is true", async () => {
+    const page = await newE2EPage();
+    // set large page to ensure test modal isn't becoming fullscreen
+    page.setViewport({ width: 1440, height: 1440 });
+    await page.setContent(`<calcite-modal style="--calcite-modal-height:600px;" fullscreen open></calcite-modal>`);
+    const modal = await page.find("calcite-modal");
+    await modal.setProperty("open", true);
+    await page.waitForChanges();
+    const style = await page.$eval("calcite-modal", (elm) => {
+      const m = elm.shadowRoot.querySelector(".modal");
+      return window.getComputedStyle(m).getPropertyValue("height");
+    });
+    expect(style).not.toEqual("600px");
+  });
+
+  it("does not overflow page bounds when requested css variable sizes are larger than viewport", async () => {
+    const page = await newE2EPage();
+    // set small page to test overflow
+    page.setViewport({ width: 800, height: 800 });
+    await page.setContent(
+      `<calcite-modal style="--calcite-modal-height:1200px;--calcite-modal-width:1200px;" open></calcite-modal>`
+    );
+    const modal = await page.find("calcite-modal");
+    await modal.setProperty("open", true);
+    await page.waitForChanges();
+    const styleW = await page.$eval("calcite-modal", (elm) => {
+      const m = elm.shadowRoot.querySelector(".modal");
+      return window.getComputedStyle(m).getPropertyValue("width");
+    });
+    const styleH = await page.$eval("calcite-modal", (elm) => {
+      const m = elm.shadowRoot.querySelector(".modal");
+      return window.getComputedStyle(m).getPropertyValue("height");
+    });
+    expect(styleW).toEqual("800px");
+    expect(styleH).toEqual("800px");
   });
 
   it("calls the beforeClose method prior to closing", async () => {