+++ src/components/tip/tip.tsx
@@ -47,7 +47,7 @@ export class Tip implements ConditionalSlotComponent, LocalizedComponent, T9nCom
   /**
    * When `true`, the component does not display.
    */
-  @Prop({ reflect: true, mutable: true }) dismissed = false;
+  @Prop({ reflect: true, mutable: true }) closed = false;
 
   /**
    * When `true`, the close button is not present on the component.
@@ -134,7 +134,7 @@ export class Tip implements ConditionalSlotComponent, LocalizedComponent, T9nCom
   // --------------------------------------------------------------------------
 
   /**
-   * Emits when the component has been dismissed.
+   * Emits when the component has been closed.
    */
   @Event({ cancelable: false }) calciteTipDismiss: EventEmitter<void>;
 
@@ -145,7 +145,7 @@ export class Tip implements ConditionalSlotComponent, LocalizedComponent, T9nCom
   // --------------------------------------------------------------------------
 
   hideTip = (): void => {
-    this.dismissed = true;
+    this.closed = true;
 
     this.calciteTipDismiss.emit();
   };